+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Added
 
+#### `compete.toml`
+
 - Added `exclude` value to `new-workspace-member`.
+- Added `manifest_dir` variable for `test-suite`.
 
 ## [0.2.2] - 2020-08-05Z